move views into a separate file
these views are probably going to get very large
add a view for /upscale (for deleting)
