+++ README.md
@@ -50,6 +50,9 @@ To generate an image from text, use the /draw command and include your prompt as
 - /stats command - shows how many /draw commands have been used.
 - /tips command - basic tips for writing prompts.
 - /upscale command - resize your image.
+- buttons - certain outputs will contain buttons.
+  - 🎲 - generates a new image with same parameters, different seed.
+  - ❌ - deletes the generated image.
 
 ## Notes
