+++ core/queuehandler.py
@@ -27,13 +27,14 @@ class DrawObject:
 
 #the queue object for extras - upscale
 class UpscaleObject:
-    def __init__(self, ctx, resize, init_image, upscaler_1, upscaler_2, upscaler_2_strength):
+    def __init__(self, ctx, resize, init_image, upscaler_1, upscaler_2, upscaler_2_strength, view):
         self.ctx = ctx
         self.resize = resize
         self.init_image = init_image
         self.upscaler_1 = upscaler_1
         self.upscaler_2 = upscaler_2
         self.upscaler_2_strength = upscaler_2_strength
+        self.view = view
 
 #the queue object for identify (interrogate)
 class IdentifyObject:
